refactor: form 타입 분리
