+++ src/hooks/api/member/useEditUserInfo.ts
@@ -2,8 +2,8 @@ import { useMutation, useQueryClient } from '@tanstack/react-query';
 import { UseFormSetError } from 'react-hook-form';
 
 import { memberApi } from '@/api/memberApis';
-import { SettingFormFields } from '@/app/(shared)/(standard)/mypage/setting/page';
 import { MEMBER_QUERY_KEYS } from '@/constants/queryKeys';
+import { SettingFormFields } from '@/types/form';
 import { setValidationError } from '@/utils/validators/setValidationError';
 
 export const useEditUserInfoMutation = (setError?: UseFormSetError<SettingFormFields>) => {
