+++ src/types/form.ts
@@ -1,0 +1,11 @@
+export interface SignupFormFields {
+  email: string;
+}
+
+export interface SettingFormFields {
+  nickname: string;
+  email: string;
+  phoneNumber?: string;
+  description?: string;
+  imageUrl?: string;
+}
